backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
coqui-tts==0.27.0

